fix: remove import cycle on Layout3D <-> HaikuElement.
